libbpf-rs: Return Result<T> from query iterators
Previously, it was impossible to tell if your iteration was terminated
b/c you ran out of "items" to look at or you hit an error. On top of
that, if the error was isolated to a single item and you still had
more items to look at, the iteration would terminate early.

This patch has the API return `Result` types which makes it possible
to report errors and also continue iteration on single item failures.

XXX: this patch has a flaw: the Iterator trait forces you to return a
`Option<Result<T>>`. This makes it impossible to return an error _and_
stop iteration. It effectively opens the user to an infinite loop if the
downstream code doesn't break out of iteration on errors.